split GeoCoord object to its own file
