+++ .gitignore
@@ -1,2 +1,4 @@
 h3/
 build/
+.cache
+*.pyc
