+++ setup.py
@@ -6,7 +6,7 @@ import os
 UBER_H3 = ('h3')
 
 
-h3py_ext = Extension('h3py', sources=['h3py.c'],
+h3py_ext = Extension('h3py', sources=['h3py.c', 'pygeocoord.c'],
                       include_dirs=[
                           join(UBER_H3, 'src', 'h3lib', 'include')
                       ],
